Avoid sets, maps of simple expressions
A `Name.Set.t` or `'a Name.Map.t` is a fine thing, but a `Simple.Set.t`
or `'a Simple.Map.t` is much sketchier, since whether we want to treat
two simples that differ only in `Rec_info.t` as equivalent depends on
context. Fortunately, we usually just need a set or map of names. The
one place where this gets tricky is in the uses of `Aliases.get_alias`,
so there's now `Aliases.Alias_set.t` to encapsulate the needed
operations.